backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0
# Pre-pinned for upcoming database and validation phases
# Database and migrations
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
